requirements.txt
fastapi[standard]
sqlalchemy
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
httpx
htmldate
pydantic
python-dotenv
asyncpg